requirements.txt
fastapi
polyline
qrcode
aiofiles
pillow
deta
